move modern FeeTable
