+++ src/screens/KeyBasedAssetTransfer/KeyBasedAssetTransferConfirm.js
@@ -30,7 +30,7 @@ import * as Table from 'components/layout/Table';
 import BalanceView from 'components/BalanceView';
 import Button from 'components/legacy/Button';
 import ContainerWithHeader from 'components/legacy/Layout/ContainerWithHeader';
-import FeeTable from 'components/other/FeeTable';
+import FeeTable from 'components/display/FeeTable';
 import Image from 'components/Image';
 import Spinner from 'components/Spinner';
 import Text from 'components/core/Text';
